requirements.txt
requests==2.32.5
tqdm==4.67.1
pandas==2.3.3
numpy==2.2.6
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
xgboost==3.2.0
nhl-api-py==3.1.1
wandb==0.24.0
python-dotenv==1.2.1
jupyterlab==4.5.1
jupyterlab_widgets==3.0.16
ipykernel==7.1.0
ipywidgets==8.1.8
redis